Added QuoteFragment to display motivational quotes on LandingActivity
