+++ app/src/main/res/layout/activity_landing.xml
@@ -95,13 +95,11 @@
             android:layout_height="wrap_content"
             android:text="Logout" />
 
-        <TextView
-            android:id="@+id/textViewQuote"
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/fragmentQuote"
+            android:name="com.example.worktracker.QuoteFragment"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="Loading quote..."
-            android:textSize="16sp"
-            android:gravity="center"
             android:layout_marginBottom="16dp" />
 
     </LinearLayout>
