+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,5 @@
     <string name="welcome">Welcome!</string>
     <string name="admin_area">Admin Area</string>
     <string name="logout">Logout</string>
+    <string name="loading_quote">Loading quote...</string>
 </resources>
